Moved all database actions into a module db.actions
