+++ openquake/engine/utils/__init__.py
@@ -16,3 +16,19 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with OpenQuake. If not, see <http://www.gnu.org/licenses/>.
 
+
+def confirm(prompt):
+    """
+    Ask for confirmation, given a ``prompt`` and return a boolean value.
+    """
+    while True:
+        try:
+            answer = raw_input(prompt)
+        except KeyboardInterrupt:
+            # the user presses ctrl+c, just say 'no'
+            return False
+        answer = answer.strip().lower()
+        if answer not in ('y', 'n'):
+            print 'Please enter y or n'
+            continue
+        return answer == 'y'
